make MacAddr struct public and not specific to linux
